Abstract different validator types into classes
